# static/openapi.yml
openapi: '3.1.0'
info:
  title: HaRSS API
  description: An RSS API in rust
  version: 0.8.0
  contact:
    name: Eric
    email: eric@pedr0.net
    url: https://pedr0.net
servers:
  - url: 'https://api.dev.harss.net/api/v1'
    description: Dev HaRSS Api, please be kind
  - url: 'http://localhost:8080/api/v1'
    description: Local HaRSS Api, for tests purpose
security:
  - basicAuth: [ ]
  - jwt: [ ]
...
paths:
...
  /user/confirm-email/{token}:
    get:
      operationId: confirm_email
      tags:
        - Users
      summary: Confirm email adress
      description: Confirm the user's email
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:    
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
  /user/update-password:
    patch:
      operationId: update_password
      tags:
        - Users
      summary: Update user's password
      description: Update user's password
      requestBody:
        required: true
        description: Update password request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
  /user/reset-password-request:
    post:
      operationId: reset_password_token
      security: []
      tags:
        - Users
      summary: Request password reset token
      description:
        Request for a password reset token that will be sent by email. The user must have registred its email
      requestBody:
        required: true
        description: Reset Password request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordTokenRequest'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
  /user/reset-password:
    post:
      operationId: reset_password
      security: []
      tags:
        - Users
      summary: Reset password
      description: Reset password using the given token
      requestBody:
        required: true
        description: Reset forgotten password request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
  /user/{userId}:
    delete:
      operationId: delete_user
      tags:
        - Users
      summary: Delete a user
      description: |
        Delete the given user.

        The caller must be admin to delete any user, or can only delete itself.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UserID'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
  /user/{userId}/update-password:
    patch:
      operationId: update_other_password
      tags:
        - Users
      summary: Update other user's password
      description: Update other user's password. Only admin can do this
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UserID'
      requestBody:
        required: true
        description: Update other's password request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOtherPasswordRequest'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
  /users:
    get:
      operationId: list_users
      tags:
        - Users
      summary: Return a list of all users
      description: |
        __Available to administrators only.__
        
        List all registered users. 
      parameters:
        - $ref: '#/components/parameters/PageSizeParameter'
        - $ref: '#/components/parameters/PageNumberParameter'
      responses:
        '200':
          description: A list of registered users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUsers'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: new_user
      tags:
        - Users
      summary: Create a new user
      description: |
        __Available to administrators only or to everyone if specified in the configuration file.__
        
        Register a new user.
      requestBody:
        required: true
        description: User creation request.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          $ref: '#/components/responses/ObjectCreatedResponse'
        default:
          $ref: '#/components/responses/Error'
  /user:
    patch:
      operationId: update_user
      tags:
        - Users
      summary: Update your user
      description: |
        Update your user.

        Currently support only email updates. An email will be sent to the new email address for confirmation.
      requestBody:
        required: true
        description: User update request.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    PageSizeParameter:
      in: query
      name: size
      required: false
      schema:
        $ref: '#/components/schemas/PageSize'
      description: Size of the page to return. Default to 20.
    PageNumberParameter:
      in: query
      name: number
      required: false
      schema:
        $ref: '#/components/schemas/PageNumber'
      description: Number of the page to return. Default to 1.
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PagedUsers:
      type: object
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required:
            - content
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/User'
...
    Page:
      type: object
      required:
        - page_number
        - page_size
        - total_items
        - total_pages
        - elements_number
      properties:
        page:
          $ref: '#/components/schemas/PageNumber'
        page_size:
          $ref: '#/components/schemas/PageSize'
        total_pages:
          type: integer
          description: Total number of pages.
        elements_number:
          type: integer
          description: Number of elements in the content
        total_items:
          type: integer
          description: Total number of elements
    PageSize:
      type: integer
      description: Size of the page to retrieve
      minimum: 1
      maximum: 200
      example: 15
      default: 20
    PageNumber:
      type: integer
      description: Number of the page to retrieve
      minimum: 1
      default: 1
      example: 1
...
    UpdatePasswordRequest:
      type: object
      description: Update password request.
      required:
        - current_password
        - new_password
        - confirm_password
      properties:
        current_password:
          $ref: '#/components/schemas/UserPassword'
        new_password:
          $ref: '#/components/schemas/UserPassword'
        confirm_password:
          $ref: '#/components/schemas/UserPassword'
    ResetPasswordTokenRequest:
      type: object
      description: A password reset token request
      required:
        - email
      properties:
        email:
          $ref: '#/components/schemas/UserEmail'
    ResetPasswordRequest:
      type: object
      description: A reset password request
      required:
        - token
        - new_password
        - confirm_password
        - login
      properties:
        username:
          $ref: '#/components/schemas/UserName'
        token:
          type: string
          format: uuid
        new_password:
          $ref: '#/components/schemas/UserPassword'
        confirm_password:
          $ref: '#/components/schemas/UserPassword'          
    UpdateOtherPasswordRequest:
      type: object
      description: Update other's password request.
      required:
        - new_password
        - confirm_password
      properties:
        new_password:
          $ref: '#/components/schemas/UserPassword'
        confirm_password:
          $ref: '#/components/schemas/UserPassword'
    UpdateUserRequest:
      type: object
      description: User update request
      properties:
        email:
          $ref: '#/components/schemas/UserEmail'
    CreateUserRequest:
      type: object
      description: User Creation request
      required:
        - username
        - password
        - confirm_password
        - role
      properties:
        username:
          $ref: '#/components/schemas/UserName'
        password:
          $ref: '#/components/schemas/UserPassword'
        confirm_password:
          $ref: '#/components/schemas/UserPassword'
        role:
          $ref: '#/components/schemas/UserRole'
        email:
          $ref: '#/components/schemas/UserEmail'
    User:
      type: object
      required:
        - id
        - username
        - role
      properties:
        id:
          $ref: '#/components/schemas/UserID'
        username:
          $ref: '#/components/schemas/UserName'
        role:
          $ref: '#/components/schemas/UserRole'
        email_verified:
          type: boolean
          description: Whether the user's email is verified, if the user has provided one.
    UserRole:
      type: string
      description: The user's role (i.e. admin or not).
      nullable: false
      minLength: 1
    UserPassword:
      type: string
      description: The user's password.
      format: password
      nullable: false
      minLength: 1
    UserID:
      type: integer
      description: The user's unique ID.
      minimum: 1
    UserName:
      type: string
      description: The user's name.
      minLength: 1
      nullable: false
    UserEmail:
      type: string
      description: The user's email.
      format: email
...
    GenericProblem:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/ProblemType'
        title:
          type: string
          description: a short description of what went wrong
        detail:
          type: string
          description: A more detailed description of the problem
        status:
          type: integer
          description: the HTTP response status
    ProblemType:
      type: string
      description: |
        a relative URI of the type of error. It can be used to identify what kind of error
        this problem is about
      example: /errors/something-bad
    ObjectCreatedId:
      type: object
      description: ID of the created object.
      properties:
        id:
          type: integer
  responses:
...
    NoContent:
      description: The request succeeded, but the response does not contain any information.
...
    ObjectCreatedResponse:
      description: The requested object has been created
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ObjectCreatedId'
    Error:
      description: Something horrible and unexpected went wrong.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericProblem'
...
    NotFound:
      description: The wanted object could not be found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericProblem'